requirements.txt
asgiref==3.9.1
django==4.2.23
django-cors-headers==4.8.0
django-filter==25.1
django-nested-admin==4.1.1
djangorestframework==3.16.1
python-dotenv==1.1.1
python-monkey-business==1.1.0
sqlparse==0.5.3
typing-extensions==4.14.1
